requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# A2A Protocol and ADK Dependencies
google-a2a>=1.0.0
google-adk>=1.0.0
uvicorn[standard]>=0.24.0
aioquic>=0.9.20
fastapi>=0.104.0
httpx>=0.25.0

# Vector Store and Search
redis>=5.0.0
sentence-transformers>=2.2.0
exa-py>=1.0.0

# RAG and LLM
google-generativeai>=0.3.0
openai>=1.0.0

# Async and Networking
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0
websockets>=12.0

# Configuration and Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Monitoring and Observability
wandb>=0.15.0
loguru>=0.7.0

# Utilities
python-dateutil>=2.8.2
requests>=2.28.0 